Added Color Picking
The colors can now be selected of what the user would time the buttons to be as they are correct, wrong, or in the worng place. You can select any color, im not sure if I will change the letters to different colors though about a memory game mode where the letters are black if you use them but it will say 3 right 2 wrong or somthing like that.